server/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-migrate
psycopg2-binary
python-dotenv
cloudinary
requests
